Wave 4 adversarial review: fix defects (E + A)
E (editor): EditorController now releases its file.watch subscription on
destruction and on file switch/re-open (~EditorController + open()-time unwatch),
and the watch callback unwatches an orphaned subscription when the controller was
destroyed or the file switched before the watch RPC resolved (SPEC 8.7 leak);
Ctrl+S is a no-op in read-only mode (SPEC 8.2). Test harness pumps the event loop
so watch subscriptions are recorded deterministically; +unwatch tests.
A (agent): rebuildRows downgrades IdleUnseen->Idle once the session is marked
seen so markSeen clears the sidebar badge (was stuck FinishedUnseen); strict
version-equality parse (rejects 1.5, matching events.ts ===); transport is a
QPointer (no UAF if the caller destroys it); +tests.

Verified: cmake build + ctest 9/9; remote typecheck + 67 tests; web typecheck.

+++ src/agent/AgentEvent.h
@@ -98,8 +98,12 @@ inline std::optional<AgentEvent> parseAgentEventLine(const QByteArray& line)
     const QJsonValue vSummary = o.value(QStringLiteral("summary"));
     const QJsonValue vMeta = o.value(QStringLiteral("metadata"));
 
-    // version === CH_EVENT_VERSION (must be the JSON number 1, not "1").
-    if (!vVersion.isDouble() || vVersion.toInt() != kAgentEventVersion)
+    // version === CH_EVENT_VERSION (must be the JSON number 1, not "1"). Compare
+    // as a double so a non-integer like 1.5 is rejected the way events.ts's
+    // strict `===` rejects it (QJsonValue::toInt() would truncate 1.5 -> 1 and
+    // wrongly accept it).
+    if (!vVersion.isDouble()
+        || vVersion.toDouble() != static_cast<double>(kAgentEventVersion))
         return std::nullopt;
     if (!vTimestamp.isString())
         return std::nullopt;
